requirements.txt
Django==2.2.2
psycopg2==2.8.3
python-dotenv==0.10.2
graphene==2.1.5
graphene-django==2.2.0
django-filter==2.0.0
django-graphql-jwt==0.2.1
django-cors-headers==3.0.2
django-environ==0.4.5
python-dateutil==2.8.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
